+++ os_in_browser/src/App.tsx
@@ -1,7 +1,7 @@
 import { ProcessDirectoryProvider } from "./context/processDirectory";
 import { FileSystemProvider } from "./context/windowFileSystem";
-import Main from "./main";
-import "./style/desktop/desktop.scss";
+import Main from "./app/Desktop";
+import "./style/desktop.scss";
 import "./style/global.scss";
 import "./style/variables.scss";
 import "./style/window.scss";
@@ -14,8 +14,7 @@ import "./style/desktop/menubar/dropdownMenu/item.scss";
 import "./style/desktop/menubar/dropdownMenu/submenu.scss";
 
 //customMenu
-import "./style/customMenu/customMenu.scss";
-import { FileDirectoryProvider } from "./context/FileDirectoryContext";
+import "./style/customMenu.scss";
 
 const App = () => {
   return (
