+++ os_in_browser/src/app/Desktop/components/dragBox/onClickDrag.tsx
@@ -1,6 +1,3 @@
-import { createSignal, JSX, onMount } from "solid-js";
-import { Dimension } from "../types/processDirectory";
-
 interface OnClickDragBoxProps {
   height: number;
   width: number;
