+++ os_in_browser/src/app/Desktop/components/menubar/icons/low-volume.png
