+++ os_in_browser/src/app/Desktop/components/window/resizer.tsx
@@ -1,35 +1,7 @@
 import { createSignal, onMount } from "solid-js";
-import useResize from "../hooks/useResize";
-import { Dimension } from "../types/processDirectory";
-
-type refType = HTMLDivElement | ((el: HTMLDivElement) => void) | undefined;
-
-interface ResizerFunctions {
-  rightBottomResize: {
-    onMouseDown: (e: MouseEvent) => void;
-  };
-  rightTopResize: {
-    onMouseDown: (e: MouseEvent) => void;
-  };
-  topResize: {
-    onMouseDown: (e: MouseEvent) => void;
-  };
-  rightResize: {
-    onMouseDown: (e: MouseEvent) => void;
-  };
-  leftBottomResize: {
-    onMouseDown: (e: MouseEvent) => void;
-  };
-  leftTopResize: {
-    onMouseDown: (e: MouseEvent) => void;
-  };
-  leftResize: {
-    onMouseDown: (e: MouseEvent) => void;
-  };
-  bottomResize: {
-    onMouseDown: (e: MouseEvent) => void;
-  };
-}
+import useResize from "../../../../hooks/useResize";
+import { Dimension } from "../../../../types/processDirectory";
+import { refType, ResizerFunctions } from "../../../../types/resizes";
 
 interface resizerProps {
   id: string;
