+++ os_in_browser/src/app/Finder/index.tsx
@@ -1,6 +1,5 @@
 import SideBar from "./sidebar";
 import FinderMain from "./finderMain";
-import { useFileSystem } from "../../context/windowFileSystem";
 
 import { FileDirectoryProvider } from "../../context/FileDirectoryContext";
