--- os_in_browser/src/components/desktop/index.tsx
@@ -1,22 +1,0 @@
-import { JSX } from "solid-js/jsx-runtime";
-import { ResolvedChildren } from "solid-js/types/reactive/signal";
-import Dock from "./dock";
-import Menubar from "./menubar";
-
-interface DesktopProps {
-  children: JSX.Element;
-}
-
-const Desktop = ({ children }: DesktopProps) => {
-  return (
-    <>
-      <Menubar />
-      <div class="mainDesktop" id="mainDesktop">
-        {children}
-        <Dock />
-      </div>
-    </>
-  );
-};
-
-export default Desktop;
