+++ os_in_browser/src/sharedComponents/FileEntry/FileInfo.tsx
