+++ os_in_browser/src/sharedComponents/FileEntry/index.tsx
