+++ os_in_browser/src/style/customMenu.scss
