+++ os_in_browser/src/style/desktop.scss
