+++ os_in_browser/src/types/customMenu.ts
@@ -1,0 +1,6 @@
+export interface customMenu {
+  title: string;
+  iconPath: string;
+  onClick: (e: MouseEvent) => void;
+  props: any; // this is for onclick Argument
+}
