+++ os_in_browser/src/types/finderSideBar.ts
@@ -1,0 +1,5 @@
+export interface Favorite {
+  src: string;
+  title: string;
+  directory: string;
+}
